Adding PRNG in LPC
This module implements a simple pseudo-random
number generator from MG based on a Galois
linear feedback shift register (GLFSR).
It can be seeded by int or string and the polynom
can be configured, although the default should be
fine.
A possible use-case is to have a reproducible
stream of pseudo-random numbers, e.g. an
individual stream per player.